style refactoring
