+++ apps/i18n/codebridge/en_us.json
@@ -7,6 +7,7 @@
   "cannotRun": "We don't know how to run your code. Please contact support.",
   "cannotStop": "We don't know how to stop your code. Please contact support.",
   "cannotTest": "We don't know how to run your tests. Please contact support.",
+  "createFileOrFolder": "Create a file or folder",
   "defaultLayout": "Default layout",
   "delete": "Delete",
   "deleteFile": "Delete file",
@@ -28,7 +29,9 @@
   "makeValidation": "Make validation file",
   "moveFile": "Move file",
   "moveFilePrompt": "Please enter your destination folder",
+  "newFile": "New File",
   "newFilePrompt": "Please name your new file",
+  "newFolder": "New Folder",
   "newFolderPrompt": "Please name your new folder",
   "noFileExtensionError": "Please include a file extension",
   "noOpenFiles": "Open a file to get started!",
