+++ apps/src/codebridge/PopUpButton/PopUpButton.tsx
@@ -4,7 +4,7 @@ import React, {useState, useCallback} from 'react';
 import Button from '@cdo/apps/componentLibrary/button';
 
 import moduleStyles from './PopUpButton.module.scss';
-import darkModeStyles from '@codebridge/styles/darkMode.module.scss';
+import darkModeStyles from '@codebridge/styles/dark-mode.module.scss';
 
 type PopUpButtonProps = {
   iconName: string;
