--- apps/src/codebridge/components/swap-layout-dropdown.module.scss
@@ -1,13 +1,0 @@
-@import 'color.scss';
-
-.layoutItem {
-  display: flex;
-  padding: 8px 16px 8px 12px;
-  align-items: center;
-  gap: 12px;
-  align-self: stretch;
-}
-
-.layoutItem:hover {
-  background-color: $neutral_dark80;
-}
