+++ apps/src/codebridge/styles/dark-mode.module.scss
@@ -11,3 +11,15 @@
     color: $light_gray_200 !important; 
   }
 }
+
+.dropdownItem {
+  display: flex;
+  padding: 8px 16px 8px 12px;
+  align-items: center;
+  gap: 12px;
+  align-self: stretch;
+
+  &:hover {
+    background-color: $neutral_dark80;
+  }
+}
